Integrate hosted gateway with group commit
