+++ exp/runtime/gateway/native_bridge.py
@@ -30,7 +30,6 @@ import threading
 import time
 from collections.abc import Callable
 from dataclasses import dataclass, field, replace
-from typing import cast
 
 from exp.common.core.artifacts import JsonObject, stable_id
 from exp.runtime.gateway.boundary import boundary_protocol_error
@@ -44,7 +43,6 @@ from exp.runtime.gateway.contracts import (
     GatewayFailure,
     GatewayFailureClass,
     GatewayRequest,
-    GatewayUsage,
 )
 from exp.runtime.gateway.discovery import (
     listing_metadata_by_alias,
@@ -60,7 +58,6 @@ from exp.runtime.gateway.execution import (
     _require_deployment_identity,  # noqa: PLC2701
 )
 from exp.runtime.gateway.group_commit import SyncGroupCommitLedger
-from exp.runtime.gateway.ledger import SQLiteAttemptLedger
 from exp.runtime.gateway.native_components import NativeGatewayComponents
 from exp.runtime.gateway.native_responses import (
     ContinuationContext,
@@ -68,6 +65,7 @@ from exp.runtime.gateway.native_responses import (
     remember_turn,
     responses_envelope,
 )
+from exp.runtime.gateway.native_settlement import terminal_from_settlement
 from exp.runtime.gateway.routing import GatewayRoute, GatewayRoutingError
 from exp.runtime.gateway.usage import GatewayUsageReport, read_usage_report, usage_html
 from exp.runtime.models.providers import (
@@ -94,12 +92,6 @@ _REQUEST_TIMEOUT_SECONDS = 120.0
 _SWEEP_GRACE_SECONDS = 5.0
 _SWEEP_INTERVAL_SECONDS = 5.0
 _SWEEP_BATCH = 16
-
-_TERMINAL_KINDS = {
-    "completed": GatewayEventKind.COMPLETED,
-    "incomplete": GatewayEventKind.INCOMPLETE,
-    "failed": GatewayEventKind.FAILED,
-}
 
 
 class _NativeDialectUnavailableError(RuntimeError):
@@ -581,7 +573,7 @@ class NativeControlPlane:
             entry = self._inflight.get(request_id)
         if entry is None:
             return "{}"
-        terminal, failure = _terminal_from_settlement(data)
+        terminal, failure = terminal_from_settlement(data)
         try:
             self._write_ledger.finish_attempt(
                 attempt_id=entry.attempt_id,
@@ -688,7 +680,7 @@ class NativeControlPlane:
             except Exception as exc:  # noqa: BLE001 - boundary sanitizes every failure.
                 raise _authority_error(exc) from exc
         return read_usage_report(
-            cast("SQLiteAttemptLedger", self._components.ledger),
+            self._components.ledger,
             organization_id=self._components.organization_id,
             identity_id=identity_id,
         )
@@ -873,7 +865,7 @@ class NativeControlPlane:
             settlement = entry.pending_settlement
             if settlement is None:
                 continue
-            terminal, failure = _terminal_from_settlement(settlement)
+            terminal, failure = terminal_from_settlement(settlement)
             self._settle_swept(request_id, entry, terminal, failure, latch_on_failure=True)
         if not abandoned:
             return
@@ -964,72 +956,3 @@ def _deployment_operation_key(route: GatewayRoute) -> str:
 def _optional_text(value: object) -> str | None:
     """Return one optional boundary string value or ``None``."""
     return value if isinstance(value, str) else None
-
-
-def _terminal_from_settlement(
-    data: JsonObject,
-) -> tuple[GatewayEvent, GatewayFailure | None]:
-    """Build the durable terminal event from one settlement payload.
-
-    Args:
-        data: Parsed settlement with ``outcome``, optional ``usage``,
-            ``tool_names``, and ``failure``.
-
-    Returns:
-        The terminal event and the optional normalized failure.
-    """
-    raw_usage = data.get("usage")
-    raw_tool_names = data.get("tool_names")
-    usage = _usage_from_payload(
-        raw_usage if isinstance(raw_usage, dict) else None,
-        [str(name) for name in raw_tool_names] if isinstance(raw_tool_names, list) else [],
-    )
-    failure_payload = data.get("failure")
-    failure = None
-    if isinstance(failure_payload, dict):
-        failure = GatewayFailure(
-            failure_class=GatewayFailureClass(str(failure_payload["failure_class"])),
-            safe_message=str(failure_payload["safe_message"]),
-        )
-    kind = _TERMINAL_KINDS[str(data["outcome"])]
-    terminal = GatewayEvent(
-        kind=kind,
-        sequence_number=0,
-        usage=usage,
-        failure=failure if kind == GatewayEventKind.FAILED else None,
-    )
-    return terminal, failure
-
-
-def _usage_from_payload(
-    payload: JsonObject | None,
-    tool_names: list[str],
-) -> GatewayUsage | None:
-    """Build normalized usage from settlement scalars.
-
-    Args:
-        payload: Optional token totals observed by the data plane.
-        tool_names: Invoked tool names in first-use order.
-
-    Returns:
-        Normalized usage, or ``None`` when nothing was observed.
-    """
-    names = tuple(str(name) for name in tool_names)
-    if payload is None or payload.get("input_tokens") is None:
-        if not names:
-            return None
-        return GatewayUsage(tool_names=names)
-    return GatewayUsage(
-        input_tokens=_optional_count(payload.get("input_tokens")),
-        output_tokens=_optional_count(payload.get("output_tokens")),
-        cached_input_tokens=_optional_count(payload.get("cached_input_tokens")),
-        reasoning_tokens=_optional_count(payload.get("reasoning_tokens")),
-        tool_names=names,
-    )
-
-
-def _optional_count(value: object) -> int | None:
-    """Return one non-negative settlement token count or ``None``."""
-    if isinstance(value, bool) or not isinstance(value, int):
-        return None
-    return value
